test: refactored requestHelpers into separate files
(cherry picked from commit d1d081562c37e1e7d9a18da01e7f3d363dabfb9c)
