+++ qa/c8-orchestration-cluster-e2e-test-suite/tests/api/v2/decision-definition/evaluate-decision-definitions-api.tests.spec.ts
@@ -21,7 +21,7 @@ import {
   EVALUATED_DECISION_EXPECTED_BODY,
   evaluateDecisionRequiredFields,
 } from '../../../../utils/beans/requestBeans';
-import {deployDecisionAndStoreResponse} from '../../../../utils/requestHelpers';
+import {deployDecisionAndStoreResponse} from '@requestHelpers';
 import {DecisionDeployment} from '@camunda8/sdk/dist/c8/lib/C8Dto';
 
 test.describe.parallel('Evaluate Decision Definitions API Tests', () => {
