+++ qa/c8-orchestration-cluster-e2e-test-suite/tests/api/v2/decision-definition/get-decision-definitions-api.tests.spec.ts
@@ -22,10 +22,11 @@ import {decisionDefinitionRequiredFields} from '../../../../utils/beans/requestB
 import {
   DECISION_DEFINITION_RESPONSE_FROM_DEPLOYMENT,
   deployDecisionAndStoreResponse,
-} from '../../../../utils/requestHelpers';
+} from '@requestHelpers';
 import {DecisionDeployment} from '@camunda8/sdk/dist/c8/lib/C8Dto';
 import fs from 'fs';
 
+/* eslint-disable playwright/expect-expect */
 test.describe.parallel('Get Decision Definitions API Tests', () => {
   const state: Record<string, unknown> = {};
   let decisionDefinition1: DecisionDeployment;
