+++ qa/c8-orchestration-cluster-e2e-test-suite/tests/api/v2/group/group-clients-api-tests.spec.ts
@@ -21,13 +21,14 @@ import {
   assignClientToGroup,
   clientIdFromState,
   createGroupAndStoreResponseFields,
-} from '../../../../utils/requestHelpers';
+} from '@requestHelpers';
 import {
   defaultAssertionOptions,
   generateUniqueId,
 } from '../../../../utils/constants';
 import {cleanupGroups} from '../../../../utils/groupsCleanup';
 
+/* eslint-disable playwright/expect-expect */
 test.describe.parallel('Groups Clients API Tests', () => {
   const state: Record<string, unknown> = {};
   state['createdIds'] = [];
