+++ qa/c8-orchestration-cluster-e2e-test-suite/tests/api/v2/group/group-users-api-tests.spec.ts
@@ -22,13 +22,14 @@ import {
   assignUsersToGroup,
   createGroupAndStoreResponseFields,
   userFromState,
-} from '../../../../utils/requestHelpers';
+} from '@requestHelpers';
 import {
   defaultAssertionOptions,
   generateUniqueId,
 } from '../../../../utils/constants';
 import {cleanupGroups} from '../../../../utils/groupsCleanup';
 
+/* eslint-disable playwright/expect-expect */
 test.describe.parallel('Group Users API Tests', () => {
   const state: Record<string, unknown> = {};
   state['createdIds'] = [];
