+++ qa/c8-orchestration-cluster-e2e-test-suite/tests/api/v2/job/job-error-api-tests.spec.ts
@@ -16,12 +16,12 @@ import {
 } from '../../../../utils/http';
 import {
   activateJobToObtainAValidJobKey,
-  setupProcessInstanceTests,
-} from '../../../../utils/requestHelpers';
+  setupProcessInstanceForTests,
+} from '@requestHelpers';
 
 /* eslint-disable playwright/expect-expect */
 test.describe('Job Error API Tests', () => {
-  const {beforeAll, beforeEach, afterEach} = setupProcessInstanceTests(
+  const {beforeAll, beforeEach, afterEach} = setupProcessInstanceForTests(
     'job_api_process',
     'jobApiProcess',
   );
