+++ qa/c8-orchestration-cluster-e2e-test-suite/tests/api/v2/resource/resource-deploy-api.spec.ts
@@ -15,17 +15,15 @@ import {
 } from '../../../../utils/http';
 import {validateResponse} from '../../../../json-body-assertions';
 import {
-  createMultiResourceFormData,
-  createResourceFormData,
-} from '../../../../utils/requestHelpers';
-import {
   validateDecisionDefinitionDeployment,
   validateDecisionRequirementsDeployment,
   validateDeploymentResponse,
   validateFormDeployment,
   validateProcessDefinitionDeployment,
   validateRpaDeployment,
-} from '../../../../utils/requestHelper/resource-deploy-requestHelper';
+  createMultiResourceFormData,
+  createResourceFormData,
+} from '@requestHelpers';
 
 /* eslint-disable playwright/expect-expect */
 test.describe.parallel('Resource Deploy API', () => {
