+++ qa/c8-orchestration-cluster-e2e-test-suite/tests/api/v2/tenant/tenant-api-tests.spec.ts
@@ -27,8 +27,9 @@ import {defaultAssertionOptions} from '../../../../utils/constants';
 import {
   assertTenantInResponse,
   createTenantAndStoreResponseFields,
-} from '../../../../utils/requestHelpers';
+} from '@requestHelpers';
 
+/* eslint-disable playwright/expect-expect */
 test.describe.parallel('Tenants API Tests', () => {
   const state: Record<string, unknown> = {};
