+++ qa/c8-orchestration-cluster-e2e-test-suite/tests/api/v2/tenant/tenant-clients-api-tests.spec.ts
@@ -24,8 +24,9 @@ import {
   assignClientsToTenant,
   clientFromState,
   createTenant,
-} from '../../../../utils/requestHelpers';
+} from '@requestHelpers';
 
+/* eslint-disable playwright/expect-expect */
 test.describe.parallel('Tenant Clients API Tests', () => {
   const state: Record<string, unknown> = {};
