+++ qa/c8-orchestration-cluster-e2e-test-suite/tests/api/v2/tenant/tenant-groups-api-tests.spec.ts
@@ -24,9 +24,10 @@ import {
   createGroupAndStoreResponseFields,
   createTenant,
   groupIdFromState,
-} from '../../../../utils/requestHelpers';
+} from '@requestHelpers';
 import {GROUPS_EXPECTED_BODY} from '../../../../utils/beans/requestBeans';
 
+/* eslint-disable playwright/expect-expect */
 test.describe.parallel('Tenant Groups API Tests', () => {
   const state: Record<string, unknown> = {};
