+++ qa/c8-orchestration-cluster-e2e-test-suite/tsconfig.json
@@ -11,7 +11,8 @@
     "paths": {
       "@fixtures": ["fixtures.ts"],
       "@setup": ["./test-setup.ts"],
-      "@pages/*": ["pages/*"]
+      "@pages/*": ["pages/*"],
+      "@requestHelpers": ["utils/requestHelpers/index.ts"]
     }
   }
 }
