+++ qa/c8-orchestration-cluster-e2e-test-suite/utils/beans/requestBeans.ts
@@ -8,21 +8,21 @@
 
 import {generateUniqueId} from '../constants';
 import * as fs from 'node:fs';
+import {readFileSync} from 'node:fs';
 import {
   createMappingRule,
-  roleDescriptionFromState,
-  roleNameFromState,
-  mappingRuleIdFromState,
-  mappingRuleNameFromState,
   mappingRuleClaimNameFromState,
   mappingRuleClaimValueFromState,
-  roleIdValueUsingKey,
-  userFromState,
   mappingRuleIdFromKey,
-} from '../requestHelpers';
+  mappingRuleIdFromState,
+  mappingRuleNameFromState,
+  roleDescriptionFromState,
+  roleIdValueUsingKey,
+  roleNameFromState,
+  userFromState,
+} from '@requestHelpers';
 import {APIRequestContext} from 'playwright-core';
 import {DecisionDeployment} from '@camunda8/sdk/dist/c8/lib/C8Dto';
-import {readFileSync} from 'node:fs';
 
 export const jobResponseFields = [
   'type',
@@ -574,4 +574,18 @@ export async function mappingRuleBundle(
 export function getExpectedContent(resourceName: string): string {
   const resourcePath = `./resources/${resourceName}`;
   return readFileSync(resourcePath, 'utf-8');
+}
+
+export function DECISION_DEFINITION_RESPONSE_FROM_DEPLOYMENT(
+  decision: DecisionDeployment,
+) {
+  return {
+    decisionDefinitionKey: decision.decisionDefinitionKey,
+    decisionRequirementsKey: decision.decisionRequirementsKey,
+    decisionRequirementsId: decision.decisionRequirementsId,
+    decisionDefinitionId: decision.decisionDefinitionId,
+    name: decision.name,
+    version: decision.version,
+    tenantId: decision.tenantId,
+  };
 }
