+++ qa/c8-orchestration-cluster-e2e-test-suite/utils/requestHelpers/authorization-requestHelpers.ts
@@ -1,0 +1,27 @@
+/*
+ * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH under
+ * one or more contributor license agreements. See the NOTICE file distributed
+ * with this work for additional information regarding copyright ownership.
+ * Licensed under the Camunda License 1.0. You may not use this file
+ * except in compliance with the Camunda License 1.0.
+ */
+
+import type {APIRequestContext} from 'playwright-core';
+import {Serializable} from 'playwright-core/types/structs';
+import {assertRequiredFields, buildUrl, jsonHeaders} from '../http';
+import {expect} from '@playwright/test';
+import {authorizedComponentRequiredFields} from '../beans/requestBeans';
+
+export async function createComponentAuthorization(
+  request: APIRequestContext,
+  body: Serializable,
+) {
+  const res = await request.post(buildUrl('/authorizations'), {
+    headers: jsonHeaders(),
+    data: body,
+  });
+
+  expect(res.status()).toBe(201);
+  const json = await res.json();
+  assertRequiredFields(json, authorizedComponentRequiredFields);
+}
